build: get version string from git
To tag a new version:

    $ git tag -a 'v0.1.19' -m 'new version v0.1.19'
    $ ./autogen.sh
    $ ./configure && make
    $ ./src/cpp/vcftools
    [...]
    Version 0.1.19

Later, be sure to push the tags to github:

    $ git push --tags

To generate a distribution tarball run:

    $ make dist
    (will create 'vcftools-0.1.19.tar.gz')

The annotated tag must start with 'v' to be properly detected by
'git-version-get' (to modify that, see --prefix parameter in the
script).

If there are commits after tag v0.1.19, the version
string will be '0.1.19.3-abcd' (where .3 is the number of commits
after the tag, and 'abcd' is the git sha1 prefix of the current
revision).

If there are uncommited files, the version string will be
'0.1.19.3-abcd-dirty' (requires running ./autogen.sh to re-generate the
version string).

Modified files:
* configure.ac: use 'git-version-get' to deduce version string based
  on wither git repository, or the '.tarball-version' file (if
  extracted from a tarball).
* Makefile.am: generate .tarball-version file in 'make dist'.
* build-aux/git-version-gen: imported script from standard GNU programs.